# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: UseCase Manager
  description: AI-powered use case extraction from workshop transcripts
  version: 0.1.0
paths:
...
  /use-cases:
    get:
      tags:
      - use-cases
      summary: List Use Cases
      description: List use cases with optional filters
      operationId: list_use_cases_use_cases_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Page Size
      - name: company_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: "null"
          title: Company Id
      - name: industry_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: "null"
          title: Industry Id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            enum:
            - new
            - under_review
            - approved
            - in_progress
            - completed
            - archived
            title: UseCaseStatus
          - type: "null"
          title: Status
      - name: assignee_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: "null"
          title: Assignee Id
      - name: min_confidence
        in: query
        required: false
        schema:
          type: number
          default: 0
          title: Min Confidence
      - name: tags
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: "null"
          title: Tags
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: "null"
          title: Q
      - name: sort_by
        in: query
        required: false
        schema:
          type: string
          default: created_at
          title: Sort By
      - name: order
        in: query
        required: false
        schema:
          type: string
          default: desc
          title: Order
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response List Use Cases Use Cases Get
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
    post:
      tags:
      - use-cases
      summary: Create Use Case
      description: Create a new use case manually
      operationId: create_use_case_use_cases_post
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
          title: Email
      - name: full_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: "null"
          title: Full Name
      - name: id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: role
        in: query
        required: true
        schema:
          type: string
          enum:
          - reader
          - maintainer
          - admin
          title: UserRole
      - name: is_active
        in: query
        required: true
        schema:
          type: boolean
          title: Is Active
      - name: is_verified
        in: query
        required: true
        schema:
          type: boolean
          title: Is Verified
      - name: created_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Created At
      - name: updated_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Updated At
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                  title: Title
                description:
                  type: string
                  title: Description
                expected_benefit:
                  anyOf:
                  - type: string
                  - type: "null"
                  title: Expected Benefit
                tags:
                  anyOf:
                  - items:
                      type: string
                    type: array
                  - type: "null"
                  title: Tags
                company_id:
                  type: string
                  format: uuid
                  title: Company Id
                transcript_id:
                  anyOf:
                  - type: string
                    format: uuid
                  - type: "null"
                  title: Transcript Id
              type: object
              required:
              - title
              - description
              - company_id
              title: UseCaseCreate
      responses:
        "201":
          description: Successful Response
          content:
            application/json:
              schema:
                properties:
                  title:
                    type: string
                    title: Title
                  description:
                    type: string
                    title: Description
                  expected_benefit:
                    anyOf:
                    - type: string
                    - type: "null"
                    title: Expected Benefit
                  tags:
                    anyOf:
                    - items:
                        type: string
                      type: array
                    - type: "null"
                    title: Tags
                  id:
                    type: string
                    format: uuid
                    title: Id
                  company_id:
                    type: string
                    format: uuid
                    title: Company Id
                  transcript_id:
                    anyOf:
                    - type: string
                      format: uuid
                    - type: "null"
                    title: Transcript Id
                  assignee_id:
                    anyOf:
                    - type: string
                      format: uuid
                    - type: "null"
                    title: Assignee Id
                  created_by_id:
                    type: string
                    format: uuid
                    title: Created By Id
                  status:
                    type: string
                    enum:
                    - new
                    - under_review
                    - approved
                    - in_progress
                    - completed
                    - archived
                    title: UseCaseStatus
                  confidence_score:
                    type: number
                    title: Confidence Score
                  effort_score:
                    anyOf:
                    - type: integer
                    - type: "null"
                    title: Effort Score
                  impact_score:
                    anyOf:
                    - type: integer
                    - type: "null"
                    title: Impact Score
                  complexity_score:
                    anyOf:
                    - type: integer
                    - type: "null"
                    title: Complexity Score
                  strategic_score:
                    anyOf:
                    - type: integer
                    - type: "null"
                    title: Strategic Score
                  priority_score:
                    anyOf:
                    - type: number
                    - type: "null"
                    title: Priority Score
                  qdrant_id:
                    anyOf:
                    - type: string
                    - type: "null"
                    title: Qdrant Id
                  created_at:
                    type: string
                    format: date-time
                    title: Created At
                  updated_at:
                    type: string
                    format: date-time
                    title: Updated At
                type: object
                required:
                - title
                - description
                - id
                - company_id
                - created_by_id
                - status
                - confidence_score
                - created_at
                - updated_at
                title: UseCaseResponse
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
...
  /use-cases/{use_case_id}/status:
    patch:
      tags:
      - use-cases
      summary: Update Status
      description: Update use case status
      operationId: update_status_use_cases__use_case_id__status_patch
      parameters:
      - name: use_case_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Use Case Id
      - name: email
        in: query
        required: true
        schema:
          type: string
          title: Email
      - name: full_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: "null"
          title: Full Name
      - name: id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: role
        in: query
        required: true
        schema:
          type: string
          enum:
          - reader
          - maintainer
          - admin
          title: UserRole
      - name: is_active
        in: query
        required: true
        schema:
          type: boolean
          title: Is Active
      - name: is_verified
        in: query
        required: true
        schema:
          type: boolean
          title: Is Verified
      - name: created_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Created At
      - name: updated_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Updated At
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                status:
                  type: string
                  enum:
                  - new
                  - under_review
                  - approved
                  - in_progress
                  - completed
                  - archived
                  title: UseCaseStatus
              type: object
              required:
              - status
              title: UseCaseStatusUpdate
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                properties:
                  title:
                    type: string
                    title: Title
                  description:
                    type: string
                    title: Description
                  expected_benefit:
                    anyOf:
                    - type: string
                    - type: "null"
                    title: Expected Benefit
                  tags:
                    anyOf:
                    - items:
                        type: string
                      type: array
                    - type: "null"
                    title: Tags
                  id:
                    type: string
                    format: uuid
                    title: Id
                  company_id:
                    type: string
                    format: uuid
                    title: Company Id
                  transcript_id:
                    anyOf:
                    - type: string
                      format: uuid
                    - type: "null"
                    title: Transcript Id
                  assignee_id:
                    anyOf:
                    - type: string
                      format: uuid
                    - type: "null"
                    title: Assignee Id
                  created_by_id:
                    type: string
                    format: uuid
                    title: Created By Id
                  status:
                    type: string
                    enum:
                    - new
                    - under_review
                    - approved
                    - in_progress
                    - completed
                    - archived
                    title: UseCaseStatus
                  confidence_score:
                    type: number
                    title: Confidence Score
                  effort_score:
                    anyOf:
                    - type: integer
                    - type: "null"
                    title: Effort Score
                  impact_score:
                    anyOf:
                    - type: integer
                    - type: "null"
                    title: Impact Score
                  complexity_score:
                    anyOf:
                    - type: integer
                    - type: "null"
                    title: Complexity Score
                  strategic_score:
                    anyOf:
                    - type: integer
                    - type: "null"
                    title: Strategic Score
                  priority_score:
                    anyOf:
                    - type: number
                    - type: "null"
                    title: Priority Score
                  qdrant_id:
                    anyOf:
                    - type: string
                    - type: "null"
                    title: Qdrant Id
                  created_at:
                    type: string
                    format: date-time
                    title: Created At
                  updated_at:
                    type: string
                    format: date-time
                    title: Updated At
                type: object
                required:
                - title
                - description
                - id
                - company_id
                - created_by_id
                - status
                - confidence_score
                - created_at
                - updated_at
                title: UseCaseResponse
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
...
  /use-cases/{use_case_id}/assignee:
    patch:
      tags:
      - use-cases
      summary: Update Assignee
      description: Assign use case to a team member
      operationId: update_assignee_use_cases__use_case_id__assignee_patch
      parameters:
      - name: use_case_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Use Case Id
      - name: email
        in: query
        required: true
        schema:
          type: string
          title: Email
      - name: full_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: "null"
          title: Full Name
      - name: id
        in: query
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: role
        in: query
        required: true
        schema:
          type: string
          enum:
          - reader
          - maintainer
          - admin
          title: UserRole
      - name: is_active
        in: query
        required: true
        schema:
          type: boolean
          title: Is Active
      - name: is_verified
        in: query
        required: true
        schema:
          type: boolean
          title: Is Verified
      - name: created_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Created At
      - name: updated_at
        in: query
        required: true
        schema:
          type: string
          format: date-time
          title: Updated At
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                assignee_id:
                  anyOf:
                  - type: string
                    format: uuid
                  - type: "null"
                  title: Assignee Id
              type: object
              title: UseCaseAssigneeUpdate
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                properties:
                  title:
                    type: string
                    title: Title
                  description:
                    type: string
                    title: Description
                  expected_benefit:
                    anyOf:
                    - type: string
                    - type: "null"
                    title: Expected Benefit
                  tags:
                    anyOf:
                    - items:
                        type: string
                      type: array
                    - type: "null"
                    title: Tags
                  id:
                    type: string
                    format: uuid
                    title: Id
                  company_id:
                    type: string
                    format: uuid
                    title: Company Id
                  transcript_id:
                    anyOf:
                    - type: string
                      format: uuid
                    - type: "null"
                    title: Transcript Id
                  assignee_id:
                    anyOf:
                    - type: string
                      format: uuid
                    - type: "null"
                    title: Assignee Id
                  created_by_id:
                    type: string
                    format: uuid
                    title: Created By Id
                  status:
                    type: string
                    enum:
                    - new
                    - under_review
                    - approved
                    - in_progress
                    - completed
                    - archived
                    title: UseCaseStatus
                  confidence_score:
                    type: number
                    title: Confidence Score
                  effort_score:
                    anyOf:
                    - type: integer
                    - type: "null"
                    title: Effort Score
                  impact_score:
                    anyOf:
                    - type: integer
                    - type: "null"
                    title: Impact Score
                  complexity_score:
                    anyOf:
                    - type: integer
                    - type: "null"
                    title: Complexity Score
                  strategic_score:
                    anyOf:
                    - type: integer
                    - type: "null"
                    title: Strategic Score
                  priority_score:
                    anyOf:
                    - type: number
                    - type: "null"
                    title: Priority Score
                  qdrant_id:
                    anyOf:
                    - type: string
                    - type: "null"
                    title: Qdrant Id
                  created_at:
                    type: string
                    format: date-time
                    title: Created At
                  updated_at:
                    type: string
                    format: date-time
                    title: Updated At
                type: object
                required:
                - title
                - description
                - id
                - company_id
                - created_by_id
                - status
                - confidence_score
                - created_at
                - updated_at
                title: UseCaseResponse
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                properties:
                  detail:
                    items:
                      properties:
                        loc:
                          items:
                            anyOf:
                            - type: string
                            - type: integer
                          type: array
                          title: Location
                        msg:
                          type: string
                          title: Message
                        type:
                          type: string
                          title: Error Type
                      type: object
                      required:
                      - loc
                      - msg
                      - type
                      title: ValidationError
                    type: array
                    title: Detail
                type: object
                title: HTTPValidationError
...
components:
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: auth/jwt/login